requirements.txt
nba_api>=1.11.0
yfpy>=17.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
tabulate>=0.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
cloudscraper>=1.2.71
